+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Logs
 
+## v0.1.0
+
+ - rewrite cli.js init: add -g/--global mode, auto-update .gitignore, create context/index.json, fix/repair symlink handling
+ - add context/project to repo
+
+
 ## v0.0.5
 
  - add context-project-guide.md